requirements.txt
apache-airflow==2.8.0
apache-airflow-providers-apache-kafka
aiohttp~=3.9.1
